Overhaul complete

+++ Tester/global.cpp
@@ -1,4 +1,4 @@
 #include "global.h"
 sf::VideoMode desktop = sf::VideoMode::getDesktopMode();
-int gd::tilesize = desktop.size.y / 10;
+float gd::tilesize = static_cast<float>(desktop.size.y / 10);
